requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
PyMySQL==1.1.0
cryptography==41.0.7
Pillow==10.2.0
bleach==6.1.0
markdown2==2.4.12
python-dotenv==1.0.0
WTForms==3.1.1
email-validator==2.1.0
gunicorn==21.2.0
